+++ src/main/kotlin/com/jetbrains/rider/aspire/manifest/ManifestService.kt
@@ -9,6 +9,7 @@ import com.intellij.notification.NotificationType
 import com.intellij.openapi.application.EDT
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
+import com.intellij.openapi.diagnostic.debug
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.LocalFileSystem
@@ -22,16 +23,27 @@ import kotlinx.coroutines.withContext
 import java.nio.file.Path
 import kotlin.io.path.absolutePathString
 
+/**
+ * A service responsible for the generation of an Aspire manifest file.
+ *
+ * This service handles the creation of an Aspire manifest file (`aspire-manifest.json`) for a given Aspire Host project.
+ */
 @Service(Service.Level.PROJECT)
 class ManifestService(private val project: Project, private val scope: CoroutineScope) {
     companion object {
         fun getInstance(project: Project) = project.service<ManifestService>()
         private val LOG = logger<ManifestService>()
 
-        private const val MANIFEST_FOLDER = "aspire-manifest"
-        private const val MANIFEST_FILE_NAME = "manifest.json"
+        private const val MANIFEST_FILE_NAME = "aspire-manifest.json"
     }
 
+    /**
+     * Generates an Aspire manifest file for the given Aspire host project.
+     *
+     * The method executes the `dotnet run --publisher manifest` command on a provided host project.
+     *
+     * @param hostProjectPath the path of the Aspire host project for which the manifest needs to be generated
+     */
     fun generateManifest(hostProjectPath: Path) {
         scope.launch(Dispatchers.Default) {
             withBackgroundProgress(project, AspireBundle.message("progress.generating.aspire.manifest")) {
@@ -43,18 +55,20 @@ class ManifestService(private val project: Project, private val scope: Coroutine
 
                 val commandLine = runtime.createCommandLine(
                     listOf(
-                        "msbuild",
-                        "/t:GenerateAspireManifest",
-                        "/p:AspireManifestPublishOutputPath=$MANIFEST_FOLDER"
+                        "run",
+                        "--publisher",
+                        "manifest",
+                        "--output-path",
+                        MANIFEST_FILE_NAME
                     )
                 )
                 val directoryPath = hostProjectPath.parent
                 commandLine.workDirectory = directoryPath.toFile()
-                LOG.debug(commandLine.commandLineString)
+                LOG.debug { commandLine.commandLineString }
 
                 val output = ExecUtil.execAndGetOutput(commandLine)
                 if (output.checkSuccess(LOG)) {
-                    val manifestPath = directoryPath.resolve(MANIFEST_FOLDER).resolve(MANIFEST_FILE_NAME)
+                    val manifestPath = directoryPath.resolve(MANIFEST_FILE_NAME)
                     val file = LocalFileSystem.getInstance().refreshAndFindFileByPath(manifestPath.absolutePathString())
                     if (file != null && file.isValid) {
                         withContext(Dispatchers.EDT) {